+++ CHANGELOG.md
@@ -26,7 +26,7 @@ Classify the change according to the following categories:
     ##### Removed
     ### Patches
 
-## Develop 2024-08-20
+## Develop 2024-09-24
 ### Major Updates
 #### Added
 - Added new model **ElectricHeaterInputs**
@@ -35,6 +35,17 @@ Classify the change according to the following categories:
 - Added new model **ASHPWaterHeaterOutputs**
 - Added new model **ASHPSpaceHeaterInputs**
 - Added new model **ASHPSpaceHeaterOutputs**
+## v3.9.4
+### Minor Updates
+#### Added
+- **portfolio_uuid** is now a field that can be added to API objects
+- **PortfolioUnlinkedRuns** tracks which run_uuids were separated from their portfolios
+- `/user/<user_uuid>/unlink_from_portfolio/` endpoint (calls `views.unlink_from_portfolio`)
+- `/summary_by_runuuids/` endpoint (calls `views.summary_by_runuuids`)
+- `/link_run_to_portfolios/` endpoint (calls `views.link_run_to_portfolios`)
+
+#### Changed
+- `UnexpectedError`, added portfolio_uuid as a field that can be returned in case of errors
 
 ## v3.9.3
 ### Minor Updates